+++ frontend/src/__test__/fixtures/geojson-test-file.geojson
@@ -1,0 +1,1 @@
+{"type":"FeatureCollection","features":[{"type":"Feature","properties":{},"geometry":{"coordinates":[[[-123.23750994470723,49.25097366768156],[-123.2376791960012,49.25087279612961],[-123.24030627043224,49.251713386099425],[-123.24079930680998,49.25154046590359],[-123.2400560728675,49.25117060678488],[-123.24037985795127,49.250992881287175],[-123.24106422187853,49.25121864059301],[-123.24135121320298,49.250978471083556],[-123.24075515429857,49.25067105240686],[-123.24113045064584,49.25033481103699],[-123.23946737271481,49.249561447195646],[-123.23973964653544,49.249306858932556],[-123.23661953572628,49.24777930178587],[-123.23592781304693,49.248461424339865],[-123.23440455140187,49.247961842547824],[-123.23517475191414,49.24679551747823],[-123.23253133053913,49.24712707866851],[-123.23241589729139,49.24841562488123],[-123.23266985043645,49.249176680239344],[-123.23413087054435,49.24996979974145],[-123.23750994470723,49.25097366768156]]],"type":"Polygon"}}]}
