+++ frontend/src/components/CreateOpeningSteps/DataForm/index.tsx
@@ -1,0 +1,16 @@
+import MapPreview from "@/components/MapPreview";
+
+
+type DataFormProps = {
+  isReview: boolean;
+}
+
+const DataForm = ({ isReview }: DataFormProps) => {
+
+  return (
+    <div>haha</div>
+  )
+
+}
+
+export default DataForm;
