+++ frontend/src/components/CreateOpeningSteps/index.tsx
@@ -1,1 +1,2 @@
 export { default as CreateOpeningFileUpload } from './FileUpload';
+export { default as CreateOpeningForm } from './DataForm';
